Move the function `read_file` from the ctl module to the SDK
