+++ changelog/+move-read-file.housekeeping.md
@@ -1,0 +1,1 @@
+Move the function `read_file` from the ctl module to the SDK.
