+++ infrahub_sdk/ctl/exceptions.py
@@ -6,9 +6,3 @@ class QueryNotFoundError(Error):
     def __init__(self, name: str, message: str = ""):
         self.message = message or f"The requested query '{name}' was not found."
         super().__init__(self.message)
-
-
-class FileNotValidError(Error):
-    def __init__(self, name: str, message: str = ""):
-        self.message = message or f"Cannot parse '{name}' content."
-        super().__init__(self.message)
